Code for Sound Chip Emulation
SN776489.ino is an arduino emulation of the sound chip as used in the Sega Game Gear
Since the Arduino has no DAC, I chose to plot the output instead. It can be viewed in the Arduino serial window, or the Python script can be used to read the serial output and put it into a CSV to be plotted.